requirements.txt
autopep8==1.6.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
cssselect==1.1.0
flake8==5.0.4
idna==2.10
lxml==4.6.3
markdownify==0.9.0
mccabe==0.7.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.5
